requirements.txt
numpy>=1.26.2
scipy>=1.11.3
matplotlib>=3.8.1
imageio>=2.33.1
pillow>=10.2.0
opencv-python>=4.9.0.80
